pyproject.toml
[project]
name = "telebot"
version = "0.0.1"
description = "Bot Telegram untuk membantu kebutuhan pengguna"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "pillow>=11.3.0",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.2",
    "yt-dlp>=2025.6.30",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.12.3",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "I" # isort
]

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"